unexport buggy function nodeSelectorAsSelector
