+++ pkg/controller/nodeipam/ipam/multi_cidr_range_allocator_test.go
@@ -35,7 +35,6 @@ import (
 	k8stesting "k8s.io/client-go/testing"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/klog/v2/ktesting"
-	v1helper "k8s.io/kubernetes/pkg/apis/core/v1/helper"
 	"k8s.io/kubernetes/pkg/controller"
 	cidrset "k8s.io/kubernetes/pkg/controller/nodeipam/ipam/multicidrset"
 	"k8s.io/kubernetes/pkg/controller/nodeipam/ipam/test"
@@ -84,7 +83,7 @@ func getTestNodeSelector(requirements []testNodeSelectorRequirement) string {
 		testNodeSelector.NodeSelectorTerms = append(testNodeSelector.NodeSelectorTerms, nst)
 	}
 
-	selector, _ := v1helper.NodeSelectorAsSelector(testNodeSelector)
+	selector, _ := nodeSelectorAsSelector(testNodeSelector)
 	return selector.String()
 }
